all: re-write using the NoVerify version of phpgrep
Also enabled named matches in -format.
This became possible as NoVerify version of phpgrep engine
has a better matching API that actually captures the matched nodes.

There is a slight performance regression due to the AST->IR conversion
that NoVerify engine does, but that's the price we have to pay.

I have an idea how to make phpgrep matching 50-200% faster,
so we may end up being even faster than before after that.

Fixes #48

Signed-off-by: Iskander Sharipov <quasilyte@gmail.com>